Migrate `Fn`-based resampling and output mechanisms to Kotlin lambdas, update serializers, tests, and migration guide accordingly.
